add lv_temp for Hk_Odd and some data type changed

+++ HK_Decode/include/HkEven.hpp
@@ -47,8 +47,8 @@ public:
     uint64_t  gps_pps_count;
     uint64_t  gps_sync_gen_count;
     uint64_t  gps_sync_send_count;
-    uint16_t  hk_head;
-    uint16_t  hk_tail;
+    uint8_t   hk_head;
+    uint8_t   hk_tail;
 
 public:
     int32_t is_bad;
